[Docs] Rename Simple* to Example* in example components
