+++ docs/src/components/examples/ExampleUseAddItemToCart.js
@@ -1,23 +1,19 @@
 import React from 'react';
 import {Button} from 'theme-ui';
-import {useAddItemsToCart, useCartCount} from 'gatsby-theme-shopify-manager';
+import {useAddItemToCart, useCartCount} from 'gatsby-theme-shopify-manager';
 import {useProducts} from '../../utils';
 
-export function SimpleUseAddItemsToCart() {
+export function ExampleUseAddItemToCart() {
   const cartCount = useCartCount();
-  const addItemsToCart = useAddItemsToCart();
+  const addItemToCart = useAddItemToCart();
   const products = useProducts();
 
   async function addToCart() {
-    const items = [
-      {
-        variantId: products[0].variants[0].shopifyId,
-        quantity: 1,
-      },
-    ];
+    const variantId = products[0].variants[0].shopifyId;
+    const quantity = 1;
 
     try {
-      await addItemsToCart(items);
+      await addItemToCart(variantId, quantity);
       alert('Successfully added that item to your cart!');
     } catch {
       alert('There was a problem adding that item to your cart.');
@@ -28,7 +24,7 @@ export function SimpleUseAddItemsToCart() {
     <>
       <p>There are currently {cartCount} items in your cart.</p>
       <Button onClick={addToCart} sx={{mb: 3}}>
-        Add items to your cart
+        Add an item to your cart
       </Button>
     </>
   );
