+++ docs/src/components/examples/ExampleUseCartCount.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {useCartCount} from 'gatsby-theme-shopify-manager';
 
-export function SimpleUseCartCount() {
+export function ExampleUseCartCount() {
   const cartCount = useCartCount();
 
   return <p>Your cart has {cartCount} items.</p>;
