+++ docs/src/components/examples/index.js
@@ -1,6 +1,6 @@
-export {SimpleUseCart} from './SimpleUseCart';
-export {SimpleUseCartItems} from './SimpleUseCartItems';
-export {SimpleUseCartCount} from './SimpleUseCartCount';
-export {SimpleUseCheckoutUrl} from './SimpleUseCheckoutUrl';
-export {SimpleUseAddItemsToCart} from './SimpleUseAddItemsToCart';
-export {SimpleUseAddItemToCart} from './SimpleUseAddItemToCart';
+export {ExampleUseCart} from './ExampleUseCart';
+export {ExampleUseCartItems} from './ExampleUseCartItems';
+export {ExampleUseCartCount} from './ExampleUseCartCount';
+export {ExampleUseCheckoutUrl} from './ExampleUseCheckoutUrl';
+export {ExampleUseAddItemsToCart} from './ExampleUseAddItemsToCart';
+export {ExampleUseAddItemToCart} from './ExampleUseAddItemToCart';
