+++ docs/src/content/hooks/useCartCount.mdx
@@ -1,16 +1,16 @@
-import {SimpleUseCartCount, ExampleWithCode} from '../../components';
+import {ExampleUseCartCount, ExampleWithCode} from '../../components';
 
 ### useCartCount()
 
 The `useCartCount()` hook provides the number of items currently in the cart. This hook returns 0 if the cart is null (and will always return a number).
 
-<ExampleWithCode element={<SimpleUseCartCount />}>
+<ExampleWithCode element={<ExampleUseCartCount />}>
 
 ```javascript
 import React from 'react';
 import {useCartCount} from 'gatsby-theme-shopify-manager';
 
-export function SimpleUseCartCount() {
+export function ExampleUseCartCount() {
   const cartCount = useCartCount();
 
   return <p>Your cart has {cartCount} items.</p>;
